[FIX] Improvements after review.
- Only load indexer when module is installed;
- Protect non standard imports;
- Standardized README.rst.
