+++ .travis.yml
@@ -35,6 +35,7 @@ install:
   - git clone --depth=1 https://github.com/OCA/maintainer-quality-tools.git ${HOME}/maintainer-quality-tools
   - export PATH=${HOME}/maintainer-quality-tools/travis:${PATH}
   - pip install --upgrade paramiko
+  - pip install --upgrade pyth
   - travis_install_nightly
 
 script:
