+++ document_rtf_index/README.rst
@@ -1,3 +1,7 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+   :alt: License: AGPL-3
+
 Index rtf documents
 ===================
 
@@ -11,6 +15,23 @@ Usage
 Just installing the module will register the rtf indexer. No further user
 action or configuration required.
 
+
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: https://runbot.odoo-community.org/runbot/knowledge/8.0
+
+.. repo_id is available in https://github.com/OCA/maintainer-tools/blob/master/tools/repos_with_ids.txt
+.. branch is "8.0" for example
+
+Bug Tracker
+===========
+
+Bugs are tracked on
+`GitHub Issues <https://github.com/OCA/knowledge/issues>`_.
+In case of trouble, please check there if your issue has already been
+reported. If you spotted it first, help us smashing it by providing a
+detailed and welcomed feedback.
+
 Credits
 =======
 
@@ -18,15 +39,15 @@ Contributors
 ------------
 
 * Ronald Portier <ronald@therp.nl>
-* Icon courtesy of http://www.picol.org (refresh.svg) and
+* Icon courtesy of https://www.picol.org (refresh.svg) and
   https://github.com/odoo/odoo/blob/8.0/addons/knowledge/static/description/icon.png
 
 Maintainer
 ----------
 
-.. image:: http://odoo-community.org/logo.png
+.. image:: https://odoo-community.org/logo.png
     :alt: Odoo Community Association
-    :target: http://odoo-community.org
+    :target: https://odoo-community.org
 
 This module is maintained by the OCA.
 
@@ -34,4 +55,4 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit http://odoo-community.org.
+To contribute to this module, please visit https://odoo-community.org.
