+++ document_rtf_index/models/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 # © 2016 Therp BV <http://therp.nl>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import models
+from . import ir_attachment
